:recycle: migrate tests about CustomView size
